# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: EDU Power Push Backend API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    API for campus electricity scanning, product user accounts, personal and
    campus statistics, push notifications, and the operator admin panel.
    Processes bind to loopback by default (Compose maps only 127.0.0.1). Public
    access is same-origin reverse proxy + TLS; operators may also use an SSH
    tunnel. Campus aggregates (`/api/v1/campus/*`) are anonymous; personal data
    uses short-lived access JWTs; admin and raw meter endpoints require an
    operator/admin session or `ADMIN_TOKEN` (never ship that token in a browser build).
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/frontend-config:
    get:
      tags: [Frontend]
      operationId: getFrontendConfig
      summary: Get non-secret product feature flags and display configuration
      security: []
      responses:
        "200":
          description: Current frontend configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FrontendConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/v1/me/channels/mail/recipient/code:
    post:
      tags: [Users]
      operationId: requestMailRecipientCode
      summary: Send a verification code before adding a notification recipient
      security:
        - userAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              additionalProperties: false
              properties:
                email: { type: string, format: email, maxLength: 254 }
      responses:
        "204":
          description: Code sent to the prospective recipient
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/me/channels/mail/recipient/verify:
    post:
      tags: [Users]
      operationId: verifyMailRecipient
      summary: Verify ownership of an additional notification recipient
      security:
        - userAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              additionalProperties: false
              properties:
                email: { type: string, format: email, maxLength: 254 }
                code: { type: string, pattern: "^[0-9]{6}$" }
      responses:
        "200":
          description: Recipient verified for the current user
          content:
            application/json:
              schema:
                type: object
                required: [email]
                additionalProperties: false
                properties:
                  email: { type: string, format: email }
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/me/channels:
    get:
      tags: [Users]
      operationId: listCurrentUserChannels
      summary: List notification channel configurations (secrets masked)
      security:
        - userAccessToken: []
      responses:
        "200":
          description: Channel configurations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Users]
      operationId: replaceCurrentUserChannels
      summary: Bulk update notification channel configurations
      security:
        - userAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelList"
      responses:
        "200":
          description: Updated configurations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/me/channels/{channel}:
    put:
      tags: [Users]
      operationId: updateCurrentUserChannel
      summary: Update one notification channel; omitted secret fields keep their stored value
      security:
        - userAccessToken: []
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelUpdate"
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/me/channels/{channel}/test:
    post:
      tags: [Users]
      operationId: testCurrentUserChannel
      summary: Send current balance, data time, location, and meter ID through a configured channel, even when delivery is disabled
      security:
        - userAccessToken: []
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      responses:
        "200":
          description: Delivery attempted; see status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelTestResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Channel is unavailable, unsupported, or has no saved configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/me/notification-settings:
    get:
      tags: [Users]
      operationId: getCurrentUserNotificationSettings
      summary: Get low-balance alert and scheduled digest rules
      security:
        - userAccessToken: []
      responses:
        "200":
          description: Notification rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Users]
      operationId: replaceCurrentUserNotificationSettings
      summary: Replace low-balance alert and scheduled digest rules
      security:
        - userAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSettings"
      responses:
        "200":
          description: Updated rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/me/push-logs:
    get:
      tags: [Users]
      operationId: listCurrentUserPushLogs
      summary: Recent push delivery history
      description: >-
        Newest first. Page through older entries with the opaque `next_cursor` from the
        previous response; a null cursor means there is nothing older.
      security:
        - userAccessToken: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 5 }
        - name: cursor
          in: query
          description: Opaque cursor returned as `next_cursor` by the previous page.
          schema: { type: string }
      responses:
        "200":
          description: Push history page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushLogPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/meters/{meter}/overview:
    get:
      tags: [Meters]
      operationId: getMeterOverview
      summary: Get latest balance and recent consumption summary for a meter
      parameters:
        - $ref: "#/components/parameters/Meter"
      responses:
        "200":
          description: Meter overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeterOverview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Internal administrator token from ADMIN_TOKEN.
    userAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived product-user access JWT.
    refreshCookie:
      type: apiKey
      in: cookie
      name: edu_power_refresh
      description: Long-lived rotating refresh JWT stored as an HttpOnly SameSite cookie.
  parameters:
...
    Meter:
      name: meter
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
...
    ChannelId:
      name: channel
      in: path
      required: true
      schema:
        type: string
        enum: [mail, sms, dingtalk, wecom, wecom_webhook, feishu, lark, discord, webhook, bark, gotify, whatsapp, pushplus, serverchan_turbo, serverchan3, mp, qq, napcat, telegram]
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid administrator token, access JWT, or refresh JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated operation is disabled by product policy
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: A required dependency is not ready
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Request rate or resource-specific cooldown exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    FrontendFeatures:
      type: object
      additionalProperties: false
      required: [auth, channels]
      properties:
        auth:
          $ref: "#/components/schemas/FrontendAuthFeatures"
        channels:
          type: object
          description: Per-channel user-interface visibility; false hides the channel and disables existing delivery.
          additionalProperties:
            type: boolean
        channel_coming_soon:
          type: object
          description: Per-channel Coming Soon state; true keeps the row visible but disables configuration and existing delivery.
          additionalProperties:
            type: boolean
        channel_order:
          type: array
          description: User-facing channel order. Missing known channels are appended automatically.
          uniqueItems: true
          items:
            type: string
            enum: [mail, sms, dingtalk, wecom, wecom_webhook, feishu, lark, discord, webhook, bark, gotify, whatsapp, pushplus, serverchan_turbo, serverchan3, mp, qq, napcat, telegram]
        channel_categories:
          type: array
          description: >-
            User-facing channel groups, rendered in this order. Each channel may belong to at most
            one category; channels left out of every category are shown in a trailing "other" group.
            Order inside a group still comes from channel_order.
          maxItems: 12
          items:
            $ref: "#/components/schemas/FrontendChannelCategory"
...
    FrontendConfig:
      type: object
      additionalProperties: false
      required: [version, features, display, updated_at]
      properties:
        version:
          type: integer
          minimum: 1
        features:
          $ref: "#/components/schemas/FrontendFeatures"
        display:
          $ref: "#/components/schemas/FrontendDisplay"
        oauth:
          $ref: "#/components/schemas/FrontendOAuthAvailability"
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    MeterOverview:
      type: object
      required: [meter, location, availability, quality]
      properties:
        meter:
          type: string
        location:
          type: object
          required: [building, floor, room]
          properties:
            building:
              type: string
            floor:
              type: string
            room:
              type: string
        latest:
          oneOf:
            - $ref: "#/components/schemas/LatestReading"
            - type: "null"
        recent_7d_kwh:
          oneOf:
            - $ref: "#/components/schemas/Decimal"
            - type: "null"
        availability:
          $ref: "#/components/schemas/Availability"
        quality:
          $ref: "#/components/schemas/Quality"
...
    Channel:
      type: object
      required: [channel, enabled, config, secret_set]
      properties:
        channel:
          type: string
          enum: [mail, sms, dingtalk, wecom, wecom_webhook, feishu, lark, discord, webhook, bark, gotify, whatsapp, pushplus, serverchan_turbo, serverchan3, mp, qq, napcat, telegram]
        enabled: { type: boolean }
        config:
          type: object
          description: Per-channel fields; secret values are masked or null on read.
          additionalProperties: true
        secret_set:
          type: object
          description: Which secret fields currently hold a stored value.
          additionalProperties: { type: boolean }
        last_result:
          oneOf:
            - $ref: "#/components/schemas/ChannelTestResult"
            - type: "null"
        updated_at: { type: string, format: date-time }
    ChannelUpdate:
      type: object
      required: [enabled, config]
      additionalProperties: false
      properties:
        enabled: { type: boolean }
        config:
          type: object
          description: Omit a secret field to keep the stored value; send null to clear it.
          additionalProperties: true
    ChannelList:
      type: object
      required: [channels]
      properties:
        channels:
          type: array
          items: { $ref: "#/components/schemas/Channel" }
    ChannelTestResult:
      type: object
      required: [status, at]
      properties:
        status: { type: string, enum: [ok, failed] }
        at: { type: string, format: date-time }
        message: { type: [string, "null"] }
        latency_ms: { type: [integer, "null"] }
    NotificationSettings:
      type: object
      required: [low_balance_alert, threshold_yuan, scheduled_digest, period, push_time]
      additionalProperties: false
      properties:
        low_balance_alert: { type: boolean }
        threshold_yuan:
          type: string
          pattern: "^(?:(?:[1-9]|[1-4][0-9])(?:\\.[0-9]{1,2})?|50(?:\\.0{1,2})?)$"
          default: "10"
          description: "Low-balance threshold in yuan, from 1 through 50"
        scheduled_digest: { type: boolean }
        period: { type: string, enum: [daily, twice, every3, weekly] }
        push_time: { type: string, pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$" }
        timezone: { type: string, default: "Asia/Shanghai" }
        templates:
          $ref: "#/components/schemas/NotificationTemplates"
          description: User-defined title and body templates applied only to non-email channels.
        updated_at: { type: string, format: date-time }
...
    PushLogPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/PushLog" }
        next_cursor: { type: [string, "null"] }
...